feat: add folders for controllers, models, repositories
