+++ recommendation-service/src/main/java/com/giftgalaxy/repositories/RecommendationRepository.java
@@ -1,5 +1,7 @@
-package com.giftgalaxy;
+package com.giftgalaxy.repositories;
 
+import com.giftgalaxy.models.Recommendation;
 import org.springframework.data.jpa.repository.JpaRepository;
+
 public interface RecommendationRepository extends JpaRepository<Recommendation, Integer> {
 }
